+++ src/screens/SearchScreen.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+const SearchScreen = ({navigation}) => {
+  return (
+    <View>
+      <Text>Some text!</Text>
+      <Text>This is where the user will search for words!</Text>
+      <Text>I dig the grey background</Text>
+    </View>
+  );
+};
+
+export default SearchScreen;
